backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic>=2.11.7
python-dotenv==1.0.1
httpx==0.28.1
hydra-db-python==0.1.6
google-genai==1.27.0
